+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support for PDF 2.0 UTF-8 strings to
   [HexaPDF::DictionaryFields::StringConverter]
+* [HexaPDF::Type::Catalog#dss] for returning the document security store
 
 ### Fixed